+++ src/bioml_workbench/__init__.py
@@ -8,15 +8,29 @@ from .configuration import (
     TrainingConfig,
     load_config,
 )
+from .configuration_manager import ConfigurationManager
 from .logging import configure_logging, get_logger
+from .utils import (
+    CacheManager,
+    FileManager,
+    SeedManager,
+    Timer,
+    timed_section,
+)
 
 __all__ = [
     "AppConfig",
+    "ConfigurationManager",
     "DashboardConfig",
     "DatasetConfig",
+    "FileManager",
+    "CacheManager",
     "LoggingConfig",
+    "SeedManager",
+    "Timer",
     "TrainingConfig",
     "configure_logging",
     "get_logger",
     "load_config",
+    "timed_section",
 ]